croco/cost-wrap: cast to int to avoid annoying warning

+++ include/proxddp/compat/crocoddyl/cost-wrap.hpp
@@ -24,12 +24,12 @@ struct CrocCostModelWrapperTpl : CostAbstractTpl<_Scalar> {
 
   /// Constructor from a crocoddyl cost model.
   explicit CrocCostModelWrapperTpl(boost::shared_ptr<CrocCostModel> cost)
-      : Base(cost->get_state()->get_ndx(), cost->get_nu()), croc_cost_(cost) {}
+      : Base((int)cost->get_state()->get_ndx(), (int)cost->get_nu()), croc_cost_(cost) {}
 
   /// Constructor using a terminal action model.
   explicit CrocCostModelWrapperTpl(
       boost::shared_ptr<crocoddyl::ActionModelAbstractTpl<Scalar>> action_model)
-      : Base(action_model->get_state()->get_ndx(), action_model->get_nu()),
+      : Base((int)action_model->get_state()->get_ndx(), (int)action_model->get_nu()),
         action_model_(action_model) {}
 
   void evaluate(const ConstVectorRef &x, const ConstVectorRef &u,
